restore+fix: 重建仓库历史并落地审查修复 P1-P9
本地 .git 目录因 git 2.55 reftable 实验特性切换 + gc 在错误目录执行而损坏。
从 origin/master (4822291, v0.30.0) 重建分支，工作区内容完整保留。

本提交涵盖 v0.30.0 之后至审查修复完成时的全部变更：
- 8c0d5e2 链：analyze_codebase 修复、重叠功能清理（automation.rs/skill_dag.rs 下线）
- P1 安装器管道死锁 -> wait_with_output
- P2 checkpoint update-ref 静默失败回读校验
- P3/P5 nanoid+esbuild 依赖漏洞升级
- P4 selection manage 竞态 OnceLock 幂等化
- P6 clippy 97 -> 0（自动+手工修复+带理由豁免）
- P9 密钥迁移失败落库 key_migration_alert + 前端 toast 警示
- vitest pool=vmThreads 兼容沙箱环境
- docs/ 归档第一阶段技术审查报告

+++ src-tauri/src/model_knowledge.rs
@@ -68,7 +68,7 @@ pub fn generation_penalty(family: &str, generation: u32) -> f32 {
         _ => return 1.0,
     };
 
-    let age = (latest as i32 - generation as i32).max(0) as u32;
+    let age = (latest - generation as i32).max(0) as u32;
     if age == 0 { return 1.0; }
     // Demote by 12% per generation, floor at 0.55
     (1.0 - 0.12 * age as f32).max(0.55)
